imflo/pyproject.toml
[tool.poetry]
name = "imflo"
version = "0.1.0"
description = ""
authors = ["Kurtis Fields <kurtisfields@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
arcade-imgui = {path = ".."}
rx = "^3.1.1"
scipy = "^1.5.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
